+++ modules/template/elements/ContainerElementVariable.php
@@ -9,17 +9,27 @@
 namespace humhub\modules\custom_pages\modules\template\elements;
 
 use humhub\libs\Html;
+use humhub\modules\custom_pages\lib\templates\twig\TwigEngine;
 use humhub\modules\custom_pages\modules\template\services\TemplateInstanceRendererService;
 use Yii;
 
+/**
+ * @property-read string $editWrapperAttributes
+ */
 class ContainerElementVariable extends BaseElementVariable
 {
-    public const EDIT_WRAPPER_ATTR_CACHE_KEY = self::class . 'editWrapperAttributes';
-
     /**
      * @var ContainerItem[]|null Cached items
      */
     private ?array $_items = null;
+
+    public function __construct(BaseElementContent $elementContent)
+    {
+        parent::__construct($elementContent);
+        TwigEngine::registerSandboxExtensionAllowedFunctions(static::class, [
+            'getEditWrapperAttributes',
+        ]);
+    }
 
     public function __toString()
     {
@@ -69,49 +79,45 @@ class ContainerElementVariable extends BaseElementVariable
             $content = Html::tag('div', Yii::t('CustomPagesModule.model', 'Empty <br />Container'));
         }
 
-        if ($this->areEditWrapperAttributesRendered()) {
+        if ($this->isEditWrapperRendered()) {
+            // Don't render the wrapper twice if the attributes are rendered
+            // in a html tag like `<div {{ container.editWrapperAttributes }}>`
             return $content;
         }
 
         $tagName = preg_match('#<(tr).+?</\1>#is', $content) ? 'tbody' : 'div';
 
-        return Html::tag($tagName, $content, $this->getEditWrapperAttributes());
+        return Html::tag($tagName, $content, $this->getEditWrapperAttributesArray());
     }
 
-    protected function getEditWrapperAttributes(): array
+    protected function getEditWrapperAttributesArray(array $attributes = []): array
     {
-        $attributes = [
-            'data-editor-container-id' => $this->elementContent->id,
-        ];
+        $attributes['data-editor-container-id'] = $this->elementContent->id;
 
         if ($this->getItems() === []) {
-            $attributes['class'] = 'cp-editor-container-empty';
+            Html::addCssClass($attributes, 'cp-editor-container-empty');
         }
 
         return $attributes;
     }
 
-    public function editWrapperAttributes(): string
+    public function getEditWrapperAttributes(array $attributes = []): string
     {
         if (TemplateInstanceRendererService::inEditMode()) {
-            $rendered = $this->getRenderedEditWrapperAttributes();
-            $rendered[] = $this->elementContent->id;
-            Yii::$app->runtimeCache->set(self::EDIT_WRAPPER_ATTR_CACHE_KEY, $rendered);
-
-            return Html::renderTagAttributes($this->getEditWrapperAttributes());
+            Yii::$app->runtimeCache->set($this->getEditWrapperAttributesCacheKey(), true);
+            $attributes = $this->getEditWrapperAttributesArray($attributes);
         }
 
-        return '';
+        return Html::renderTagAttributes($attributes);
     }
 
-    private function areEditWrapperAttributesRendered(): bool
+    private function isEditWrapperRendered(): bool
     {
-        return in_array($this->elementContent->id, $this->getRenderedEditWrapperAttributes());
+        return Yii::$app->runtimeCache->get($this->getEditWrapperAttributesCacheKey()) === true;
     }
 
-    private function getRenderedEditWrapperAttributes(): array
+    private function getEditWrapperAttributesCacheKey(): string
     {
-        $rendered = Yii::$app->runtimeCache->get(self::EDIT_WRAPPER_ATTR_CACHE_KEY);
-        return is_array($rendered) ? $rendered : [];
+        return 'customPageTemplateContainerEditWrapperAttributes' . $this->elementContent->id;
     }
 }